flask/requirements.txt
aniso8601==9.0.1
beautifulsoup4==4.10.0
black==22.1.0
boto3==1.21.19
botocore==1.24.19
certifi==2021.10.8
chardet==4.0.0
click==8.0.4
et-xmlfile==1.1.0
Flask==2.0.3
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
googlemaps==4.6.0
gunicorn==20.1.0
idna==2.10
itsdangerous==2.1.1
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.1.0
marshmallow==3.15.0
mypy-extensions==0.4.3
numpy==1.23.0
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
pathspec==0.9.0
phonenumbers==8.12.45
platformdirs==2.5.1
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.25.1
s3transfer==0.5.2
six==1.16.0
soupsieve==2.3.1
tomli==2.0.1
typing_extensions==4.1.1
uk-postcode-utils==1.1
urllib3==1.26.9
Werkzeug==2.0.3